+++ bindingcollectionadapter-recyclerview/src/androidTest/java/me/tatarka/bindingcollectionadapter2/recyclerview/DynamicItemViewTest.java
@@ -10,6 +10,7 @@ import org.junit.runner.RunWith;
 import java.util.Arrays;
 import java.util.List;
 
+import androidx.annotation.NonNull;
 import androidx.databinding.DataBindingUtil;
 import androidx.databinding.ObservableArrayList;
 import androidx.databinding.ObservableList;
@@ -91,7 +92,7 @@ public class DynamicItemViewTest {
         items.addAll(Arrays.asList("one", "two", "three"));
         ItemBinding<String> itemBinding = ItemBinding.of(new OnItemBind<String>() {
             @Override
-            public void onItemBind(ItemBinding itemBinding, int position, String item) {
+            public void onItemBind(@NonNull ItemBinding itemBinding, int position, String item) {
                 itemBinding.set(BR.item, position);
             }
         });